Time: 5 ms (100.00%), Space: 7.8 MB (40.00%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.cpp
@@ -3,7 +3,7 @@ public:
     int scoreOfString(string s) {
         int sum=0;
         for(int i=0;i<s.size()-1;i++){
-            sum+=abs(static_cast<int>(s[i]) - static_cast<int>(s[i + 1]));
+            sum += abs(s[i] - s[i + 1]);
         }
         return sum;
     }
